requirements.txt
pandas
streamlit
feedparser
nltk
requests
python-dateutil